Bugfixes
Fixed a bug in supervised classification where first point of new poly was on top left corner of grid.
Fixed a bug where only the edge of a polygon was used in class definitions
reverted graph tool to plotting maps with matplotlib library instead of modestimage, because of bugs in zooming.